+++ packages/terra-slide-panel/src/SlidePanel.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './SlidePanel.scss';
+import styles from './SlidePanel.module.scss';
 
 const cx = classNames.bind(styles);